1st modification

+++ add.c
@@ -1,6 +1,6 @@
 #include <arithmetic.h>
 
-int add(int a, int b)
+float add(float a, float b)
 {
     return a + b;
 }
